% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.pcs.R
\name{as.pcs}
\alias{as.pcs}
\title{Converts Parameter to PCS}
\usage{
as.pcs(obj)
}
\arguments{
\item{obj}{[\code{ParamSet|Param}]}
}
\description{
Coverts a ParamSet or a Param form ParamHelpers to lines of PCS definitions.
}
