% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmac.R
\name{rsmac}
\alias{rsmac}
\title{Starts SMAC}
\usage{
rsmac(fun, scenario, params = NULL, path.to.smac = "~/bin/smac",
  cl.args = list(), id.smac.run = NULL, cleanup = TRUE, par.id = 1,
  start.timeout = 30)
}
\arguments{
\item{fun}{[\code{smoof_function}]}

\item{scenario}{[\code{list}] \cr
scenario description}

\item{params}{[\code{character}] \cr
Params in pcs format. Default is \code{\link{as.pcs}(getParamSet(fun))}.}

\item{path.to.smac}{[\code{character(1)}] \cr
The directory where the smac binary is located.
All intermediate files will be saved there.}

\item{cl.args}{[\code{list}] \cr
CL arguments that should be passed when calling \code{./smac}.}

\item{id.smac.run}{[\code{character(1)}] \cr
You might want to overrride the automaticly generated ID to do restarts or parallel optimizations on the exact same scenario.}

\item{cleanup}{[\code{logical(1)}] \cr
Should all files be deleted after the run?}

\item{par.id}{[\code{integer(1)}] \cr
For parallel usage on the same id.smac.run this helps to specify the which parallel run we are in.}

\item{start.timeout}{[\code{integer(1)}] \cr
For parallel usage on the same id.smac.run this helps to specify the which parallel run we are in.}
}
\value{
\link[ParamHelpers]{OptPath}
}
\description{
Starts SMAC on the given function, with the given scenario and the given params.
}
\examples{
 \dontrun{
 scenario = list("use-instances" = "false", runObj = "QUALITY", numberOfRunsLimit = 5)
 res = rsmac(makeBraninFunction(), scenario = scenario)
 best.idx = getOptPathBestIndex(res)
 getOptPathEl(res, best.idx)
 as.data.frame(res)
 }
}
