% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rsmac_minimize}
\alias{rsmac_minimize}
\title{Smac minimization function}
\usage{
rsmac_minimize(grid, objective, pysmac_args = NULL, init_rcode = NULL)
}
\arguments{
\item{grid}{list like list(x1=list(type='continuous', init=0, min=-5, max=10), x2=..)}

\item{objective}{objective function to minimize}

\item{pysmac_args}{list of pysmac additional parameters. The description is here:
\url{
https://github.com/automl/pysmac/blob/a3452d56aa1f3352c36ec0750be75a1f8fafe509/
pysmac/optimize.py#L32-L38
}}

\item{init_rcode}{r code expression that will be runned once before pysmac}
}
\description{
Smac minimization function
}
\examples{
\dontrun{
objective <- function(x1, x2) {
  (x2 - (5.1 / (4*pi^2))*x1^2 + (5 / pi)*x1 - 6)^2 +
    10*(1-(1 / (8*pi))) * cos(x1) + 10
}

grid <- list(
  x1=list(type='continuous', init=0, min=-5, max=10),
  x2=list(type='continuous', init=0, min=0, max=15))

res <- rsmac_minimize(grid, objective, list(max_evaluations=100))

stopifnot(abs(res$target_min - 0.6) < 0.1)
stopifnot(abs(res$optimized_x$x1 - 2.937) < 0.1)
stopifnot(abs(res$optimized_x$x2 - 2.365) < 0.1)
cat('\\n')
print(res) 
}
}

