% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rsmacMinimize}
\alias{rsmacMinimize}
\title{Smac minimization function}
\usage{
rsmacMinimize(objective, grid, ...)
}
\arguments{
\item{objective}{objective function to minimize}

\item{grid}{list like list(x1=list(type='continuous', init=0, min=-5, max=10), x2=..)}

\item{...}{additional parameters. The description is right after 'x_categorical' here:
https://github.com/automl/pysmac/blob/
a3452d56aa1f3352c36ec0750be75a1f8fafe509/pysmac/optimize.py#L28}
}
\description{
Smac minimization function
}
\examples{
\dontrun{
objective <- function(x1, x2) {
  (x2 - (5.1 / (4*pi^2))*x1^2 + (5 / pi)*x1 - 6)^2 +
    10*(1-(1 / (8*pi))) * cos(x1) + 10
}

grid <- list(
  x1=list(type='continuous', init=0, min=-5, max=10),
  x2=list(type='continuous', init=0, min=0, max=15))

res <- rsmacMinimize(objective, grid, max_evaluations=15)
cat('\\n')
print(res) 
}
}

